test: address multipart version PR review feedback
Make the 0.0.100 regression test fail on master by removing multipart
fallback, and move _version_str_to_tuple tests to test_dependencies_utils.
